fix: setting validation off by default, tests refactor
