+++ src/BatchFirestoreRepository.ts
@@ -1,7 +1,7 @@
 import { CollectionReference, WriteBatch } from '@google-cloud/firestore';
-import { IEntity, WithOptionalId } from './types';
-import { serializeEntity, Instantiable } from '.';
+import { IEntity, WithOptionalId, Instantiable } from './types';
 import { getMetadataStorage, CollectionMetadata } from './MetadataStorage';
+import { serializeEntity } from './utils';
 
 // TODO: Eventually I'll abstract this class since right
 // now is tied to WriteBatch class from firestore
