+++ src/helpers.ts
@@ -112,22 +112,3 @@ export const createBatch = () => {
 
   return new FirestoreBatch(null);
 };
-
-/**
- * Returns a serializable object from entity<T>
- *
- * @template T
- * @param {T} Entity object
- * @param {CollectionMetadata[]} subColMetadata Subcollection
- * metadata to remove runtime-created fields
- * @returns {Object} Serialiable object
- */
-export function serializeEntity<T extends IEntity>(
-  obj: T,
-  subColMetadata: CollectionMetadata[]
-): Object {
-  subColMetadata.forEach(scm => {
-    delete obj[scm.propertyKey];
-  });
-  return { ...obj };
-}
